Refactored code for corresponding system design

+++ src/main/java/com/pgl1/database/repository/OrderHistoryRepository.java
@@ -1,10 +1,10 @@
 package com.pgl1.database.repository;
 
-import com.pgl1.database.model.entity.OrderHistory;
 import org.springframework.data.jpa.repository.JpaRepository;
+import com.pgl1.database.model.entity.OrderHistory;
 
 import java.util.List;
 
 public interface OrderHistoryRepository extends JpaRepository<OrderHistory, Integer> {
-    List<OrderHistory> findByOrderHistoryId(Integer orderHistoryId);
+    List<OrderHistory> findByOrderHistoryId(Long orderHistoryId);
 }
