+++ src/main/java/com/pgl1/database/repository/PickupPointRepository.java
@@ -1,10 +1,10 @@
 package com.pgl1.database.repository;
 
-import com.pgl1.database.model.entity.PickupPoint;
 import org.springframework.data.jpa.repository.JpaRepository;
+import com.pgl1.database.model.entity.PickupPoint;
 
 import java.util.List;
 
 public interface PickupPointRepository extends JpaRepository<PickupPoint, Integer> {
-    List<PickupPoint> findByPickupPointId(Integer pickupPointId);
+    List<PickupPoint> findByPickupPointId(Long pickupPointId);
 }
